requirements.txt
aiosignal==1.3.2
gunicorn==20.1.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
blosc2==3.1.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Deprecated==1.2.18
et_xmlfile==2.0.0
Flask==3.1.0
flask-ngrok==0.0.25
future==1.0.0
geographiclib==2.0
geopy==2.4.1
googleapis-common-protos==1.59.1
googlemaps==4.10.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
line-bot-sdk==2.4.1
MarkupSafe==3.0.2
matplotlib==3.8.0
msgpack==1.1.0
ndindex==1.9.2
numexpr==2.10.2
numpy==1.25.2
OpenCC==1.1.9
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.31.1
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-flask==0.52b1
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-wsgi==0.52b1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
packaging==24.2
pandas==2.2.3
prometheus_client==0.21.1
propcache==0.3.0
protobuf==4.25.6
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
PyMySQL==1.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.29.0
six==1.17.0
sortedcontainers==2.4.0
tables==3.10.2
threadpoolctl==3.5.0
thrift==0.21.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xgboost==2.1.4
zipp==3.21.0
zope.interface==7.2
